requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
langsmith
beautifulsoup4
streamlit